refactor: ContentPage 로직 분리
