+++ src/pages/NotFound/index.ts
@@ -2,16 +2,12 @@ import View from '~/core/view'
 
 const template = () => `<h1>404 Not Found</h1>`
 
-interface State {}
-
 interface Props {
   parentId: string
 }
 
-const initialState: State = {}
-
-export default class NotFoundPage extends View<State> {
+export default class NotFoundPage extends View {
   constructor({ parentId }: Props) {
-    super({ parentId, initialState, template })
+    super({ parentId, template })
   }
 }
